Borogove5: experiment on gravity, change shared memory and run with runPycuda gravity.py -b <blockSize>, see results

+++ trunk/anuga_cuda/gravity/gravity.cu
@@ -1,5 +1,5 @@
 __global__ void gravity_wb(
-        long N,
+        int N,
         double g,
         double * stage_vertex_values, 
         double * stage_edge_values,
